Correct source file references
It appears that the source code files have changed since this section was last edited. Consequently, the included source snippets are wrong and the end of this section does not make sense.
http://guides.grails.org/creating-your-first-grails-app/guide/index.html#services
Also, one of the include directives referred to a scaffold controller which does not declare a service.
This commit fixes these problem, at least until the groovy source files are inadvertently changed again!